HBASE-17534 Avoid re-wrapping IOExceptions as IOExceptions
This ensures that when a DoNotRetryIOException is returned
from the RPC, it is propagated as a DoNotRetryIOException.
When this does not happen, requests which should not be
retried are repeated per the client configuration.

(cherry picked from commit afb67a8d9daf9dacce580067e6dc96c8d8956a0e)

Change-Id: I92c48542ec86afef983bf6ff57a3e68a8dd74b29
Author: Josh Elser
Reason: Bug
Ref: CDH-57968