HBASE-21905: [FIFOCompaction] Don't compact already inflight store files (#1829)
This one surfaced as a flake test but turns out to be a legit bug
in FIFOCompaction code. FifoCompaction does not check if an empty
store file is already being compacted by an in-flight compaction
request and still enqueues. It can potentially race with a running
compaction (as in this test case, see jira for the exact exception).

Fixes the edge case and cleans up the test code a bit.

Signed-off-by: Andrew Purtell <apurtell@apache.org>
(cherry picked from commit 92bfd928b570c1161efde89609ace463f4d177c6)

Change-Id: Ib23320b8570d635ae461f5f6fe4b382b94013bc0