HBASE-25692 Always try to close the WAL reader when we catch any exception (#3090)
There are code paths in which we throw non-IOExceptions when
initializing a WAL reader. However, we only close the InputStream to the
WAL filesystem when the exception is an IOException. Close it if it is
open in all cases.

Co-authored-by: Josh Elser <jelser@cloudera.com>
Signed-off-by: Andrew Purtell <apurtell@apache.org>
(cherry picked from commit a227a4aa371b2a591c4bf9f92a27d2cdcf749eae)
Change-Id: Id705661faa3f8eb3fcec59ddaef5fcb3f0138f87